requirements.txt
pandas==2.0.3
numpy==1.24.4
requests==2.28.1
tqdm==4.64.1
scipy==1.9.1
lightgbm==4.6.0
xgboost==2.1.3
scikit-learn==1.0.2
pyyaml==6.0